[WRFE-15](rename): form 파일 위치 변경
