+++ apps/front/wraffle-webview/src/entities/product/model/index.ts
