+++ apps/front/wraffle-webview/src/entities/product/ui/ThumbnailCard.tsx
