+++ apps/front/wraffle-webview/src/entities/product/ui/index.ts
@@ -1,0 +1,1 @@
+export {ThumbnailCard} from './ThumbnailCard';
